requirements.txt
SQLAlchemy~=2.0.26
requests~=2.31.0
pydantic~=2.5.0
plotly~=5.19.0
fastapi~=0.109.2
starlette~=0.36.3
uvicorn~=0.27.1
aiogram~=3.4.1
kaleido==0.2.1
python-dotenv~=1.0.1
psycopg2